misc: Rename petsc/utils.py
