+++ devito/petsc/internals.py
@@ -2,10 +2,7 @@ import os
 import ctypes
 from pathlib import Path
 
-from devito.tools import memoized_func, filter_ordered, as_tuple
-from devito.types import Symbol, SteppingDimension, TimeDimension
-from devito.operations.solve import eval_time_derivatives
-from devito.symbolics import retrieve_functions, retrieve_dimensions
+from devito.tools import memoized_func
 
 
 class PetscOSError(OSError):
@@ -106,43 +103,3 @@ petsc_type_mappings, petsc_type_to_ctype = get_petsc_type_mappings()
 
 
 petsc_languages = ['petsc']
-
-
-def get_funcs(exprs):
-    funcs = [
-        f for e in exprs
-        for f in retrieve_functions(eval_time_derivatives(e.lhs - e.rhs))
-    ]
-    return as_tuple(filter_ordered(funcs))
-
-
-def generate_time_mapper(exprs):
-    """
-    Replace time indices with `Symbols` in expressions used within
-    PETSc callback functions. These symbols are Uxreplaced at the IET
-    level to align with the `TimeDimension` and `ModuloDimension` objects
-    present in the initial lowering.
-    NOTE: All functions used in PETSc callback functions are attached to
-    the `SolverMetaData` object, which is passed through the initial lowering
-    (and subsequently dropped and replaced with calls to run the solver).
-    Therefore, the appropriate time loop will always be correctly generated inside
-    the main kernel.
-    Examples
-    --------
-    >>> exprs = (Eq(f1(t + dt, x, y), g1(t + dt, x, y) + g2(t, x, y)*f1(t, x, y)),)
-    >>> generate_time_mapper(exprs)
-    {t + dt: tau0, t: tau1}
-    """
-    # First, map any actual TimeDimensions
-    time_indices = [d for d in retrieve_dimensions(exprs) if isinstance(d, TimeDimension)]
-
-    funcs = get_funcs(exprs)
-
-    time_indices.extend(list({
-        i if isinstance(d, SteppingDimension) else d
-        for f in funcs
-        for i, d in zip(f.indices, f.dimensions)
-        if d.is_Time
-    }))
-    tau_symbs = [Symbol('tau%d' % i) for i in range(len(time_indices))]
-    return dict(zip(time_indices, tau_symbs))
