+++ devito/petsc/types/metadata.py
@@ -8,7 +8,7 @@ from devito.symbolics.extraction import separate_eqn, generate_targets, centre_s
 from devito.types.equation import Eq
 from devito.operations.solve import eval_time_derivatives
 
-from devito.petsc.utils import petsc_variables
+from devito.petsc.internals import petsc_variables
 from devito.petsc.types.equation import EssentialBC, ZeroRow, ZeroColumn
